move chart components out
